+++ client/src/test/java/replicant/ZoneTest.java
@@ -12,13 +12,13 @@ public class ZoneTest extends AbstractReplicantTest {
 
         final Zone zone1 = Replicant.createZone();
 
-        assertEquals(ReplicantZoneHolder.getDefaultZone().getContext(), Replicant.context());
+        assertEquals(ReplicantZoneHolder.getDefaultZone().getReplicantContext(), Replicant.context());
         assertEquals(ReplicantZoneHolder.getZoneStack().size(), 0);
         assertFalse(zone1.isActive());
 
         final String expected = ValueUtil.randomString();
         final String actual = zone1.safeRun(() -> {
-            assertEquals(zone1.getContext(), Replicant.context());
+            assertEquals(zone1.getReplicantContext(), Replicant.context());
             assertEquals(ReplicantZoneHolder.getZoneStack().size(), 1);
             assertTrue(zone1.isActive());
             return expected;
@@ -26,7 +26,7 @@ public class ZoneTest extends AbstractReplicantTest {
 
         assertEquals(actual, expected);
 
-        assertEquals(ReplicantZoneHolder.getDefaultZone().getContext(), Replicant.context());
+        assertEquals(ReplicantZoneHolder.getDefaultZone().getReplicantContext(), Replicant.context());
         assertEquals(ReplicantZoneHolder.getZoneStack().size(), 0);
     }
 
@@ -36,17 +36,17 @@ public class ZoneTest extends AbstractReplicantTest {
 
         final Zone zone1 = Replicant.createZone();
 
-        assertEquals(ReplicantZoneHolder.getDefaultZone().getContext(), Replicant.context());
+        assertEquals(ReplicantZoneHolder.getDefaultZone().getReplicantContext(), Replicant.context());
         assertEquals(ReplicantZoneHolder.getZoneStack().size(), 0);
         assertFalse(zone1.isActive());
 
         zone1.safeRun(() -> {
-            assertEquals(zone1.getContext(), Replicant.context());
+            assertEquals(zone1.getReplicantContext(), Replicant.context());
             assertEquals(ReplicantZoneHolder.getZoneStack().size(), 1);
             assertTrue(zone1.isActive());
         });
 
-        assertEquals(ReplicantZoneHolder.getDefaultZone().getContext(), Replicant.context());
+        assertEquals(ReplicantZoneHolder.getDefaultZone().getReplicantContext(), Replicant.context());
         assertEquals(ReplicantZoneHolder.getZoneStack().size(), 0);
     }
 
@@ -56,20 +56,20 @@ public class ZoneTest extends AbstractReplicantTest {
 
         final Zone zone1 = Replicant.createZone();
 
-        assertEquals(ReplicantZoneHolder.getDefaultZone().getContext(), Replicant.context());
+        assertEquals(ReplicantZoneHolder.getDefaultZone().getReplicantContext(), Replicant.context());
         assertEquals(ReplicantZoneHolder.getZoneStack().size(), 0);
         assertFalse(zone1.isActive());
 
         assertThrows(
                 ParseException.class,
                 () -> zone1.run(() -> {
-                    assertEquals(zone1.getContext(), Replicant.context());
+                    assertEquals(zone1.getReplicantContext(), Replicant.context());
                     assertEquals(ReplicantZoneHolder.getZoneStack().size(), 1);
                     assertTrue(zone1.isActive());
                     throw new ParseException("", 1);
                 }));
 
-        assertEquals(ReplicantZoneHolder.getDefaultZone().getContext(), Replicant.context());
+        assertEquals(ReplicantZoneHolder.getDefaultZone().getReplicantContext(), Replicant.context());
         assertEquals(ReplicantZoneHolder.getZoneStack().size(), 0);
     }
 
@@ -79,19 +79,19 @@ public class ZoneTest extends AbstractReplicantTest {
 
         final Zone zone1 = Replicant.createZone();
 
-        assertEquals(ReplicantZoneHolder.getDefaultZone().getContext(), Replicant.context());
+        assertEquals(ReplicantZoneHolder.getDefaultZone().getReplicantContext(), Replicant.context());
         assertEquals(ReplicantZoneHolder.getZoneStack().size(), 0);
         assertFalse(zone1.isActive());
 
         final Procedure procedure = () -> {
-            assertEquals(zone1.getContext(), Replicant.context());
+            assertEquals(zone1.getReplicantContext(), Replicant.context());
             assertEquals(ReplicantZoneHolder.getZoneStack().size(), 1);
             assertTrue(zone1.isActive());
             throw new ParseException("", 1);
         };
         assertThrows(ParseException.class, () -> zone1.run(procedure));
 
-        assertEquals(ReplicantZoneHolder.getDefaultZone().getContext(), Replicant.context());
+        assertEquals(ReplicantZoneHolder.getDefaultZone().getReplicantContext(), Replicant.context());
         assertEquals(ReplicantZoneHolder.getZoneStack().size(), 0);
     }
 
@@ -101,18 +101,18 @@ public class ZoneTest extends AbstractReplicantTest {
 
         final Zone zone1 = Replicant.createZone();
 
-        assertEquals(ReplicantZoneHolder.getDefaultZone().getContext(), Replicant.context());
+        assertEquals(ReplicantZoneHolder.getDefaultZone().getReplicantContext(), Replicant.context());
         assertEquals(ReplicantZoneHolder.getZoneStack().size(), 0);
         assertFalse(zone1.isActive());
 
         final Procedure procedure = () -> {
-            assertEquals(zone1.getContext(), Replicant.context());
+            assertEquals(zone1.getReplicantContext(), Replicant.context());
             assertEquals(ReplicantZoneHolder.getZoneStack().size(), 1);
             assertTrue(zone1.isActive());
         };
         zone1.run(procedure);
 
-        assertEquals(ReplicantZoneHolder.getDefaultZone().getContext(), Replicant.context());
+        assertEquals(ReplicantZoneHolder.getDefaultZone().getReplicantContext(), Replicant.context());
         assertEquals(ReplicantZoneHolder.getZoneStack().size(), 0);
     }
 }