+++ server/src/main/java/replicant/server/ServerConstants.java
@@ -5,17 +5,17 @@ package replicant.server;
  */
 public final class ServerConstants {
     /**
-     * Key added to the context when passing through replication method.
-     * Used to ensure that there can be at most one replication context active.
+     * Key added to the request context when passing through a replication method.
+     * Used to ensure that there can be at most one replication invocation active.
      */
     public static final String REPLICATION_INVOCATION_KEY = "ReplicationActive";
     /**
-     * Key used to retrieve an opaque identifier for the session from the ReplicantContextHolder.
+     * Key used to retrieve an opaque identifier for the session from the ReplicantRequestContextHolder.
      * Used to pass data from the servlet to the EJB.
      */
     public static final String SESSION_ID_KEY = "SessionID";
     /**
-     * Key used to retrieve an opaque identifier for the request from the ReplicantContextHolder.
+     * Key used to retrieve an opaque identifier for the request from the ReplicantRequestContextHolder.
      * Used to pass data from the servlet to the EJB.
      */
     public static final String REQUEST_ID_KEY = "RequestID";